review fixes: put the append arithmetic where it is used, name the underscore rule
Standards review found _appended_duration and _track_end sitting in media.py
with zero callers there — both are timeline frame math (one is how many frames
a clip occupies once appended, the other where a track ends), and pool.py was
their only consumer. Moved to timeline.py, which also makes media.py's "clips,
bins, and the helpers that build clips from paths" docstring true.

The same finding asked why a sibling imports underscore-prefixed names at all.
It is deliberate, so __init__.py now says so: inside this package a leading
underscore means "__init__ does not re-export this", which is what keeps the
public surface identical to the pre-split fakes.py. Making them public instead
would widen that surface, which the ticket's parity goal rules out.

Also rewrote a sentence in the package docstring that did not parse.
